requirements.txt
openai-whisper
discord
py-cord==2.3.2
python-dotenv